Extensive changes--read notes:
Bubbles and 2D Net were competing to be the view that got to assign the x and y attributes on the original data node. This was undesirable, as it led to general instability. This had also happened before, only between the 2d and 3d networks. To solve it then, I used UltraDeepClone to make copies of the data for 3d net, and let 2d net continue to hang its own properties onto the original object.
No more. Now, there are functions inside of app that will hand you either references to the original objects or copies of them. I'm certain this has broken many things, but hopefully these new bugs will be buried deep in the interactive features of the app. We'll see.