Removed copyright notice
I did notice that you have a copyright notice in your readme. As a work of CDC, we don’t actually have a copyright, so our material is not copyright CDC and this makes the copyright notice line probably inaccurate. The next section describes the public domain /cc1 release and is more detailed, and in conflict with the copyright line.